+++ CHANGELOG.md
@@ -4,8 +4,24 @@ All notable changes to this project will be documented in this file. See [standa
 
 ## Features
 
-#12, #13 Supporting more input parameters
-#17 Use stack name to filter for current stack
-#23 Output to GitHub actions
+### 1.6.1 , 1.6.2
+
+- Change log updates
+
+### 1.6.0
+
+- Added support for change set description
+- Added support for YAML output format
+- Added Retry for Rate Limit Exception
+
+### 1.5.0
+
+- Added support for YAML input format
+
+### older versions
+
+- #12, #13 Supporting more input parameters
+- #17 Use stack name to filter for current stack
+- #23 Output to GitHub actions
 
 ## Bug Fixes